[modbus] Remove abstraction (unnecessary interfaces not meant for exten)
